# Speed up CLI startup by ~35% by lazy-loading `DOMParser`