## Type of change

<!-- (mark with an `X`) -->

```
- [ ] Bug fix
- [ ] New feature development
- [x] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [ ] Build/deploy pipeline (DevOps)
- [ ] Other
```

## Objective

<!--Describe what the purpose of this PR is. For example: what bug you're fixing or what new feature you're adding-->

Bitwarden's CLI takes ~2.5s to run `bw --version` on my laptop. When poking around in `bw.ts`, I noticed that an instance of `JSDOM` was created _every_ time the CLI starts up, even though it's rarely used. If that polyfill is lazily loaded, the CLI starts up ~500ms faster in **all circumstances**. When running `bw --version`, thats ~37% faster!

```
Benchmark 1: node new/bw.js --version
  Time (mean ± σ):      1.772 s ±  0.018 s    [User: 1.861 s, System: 0.212 s]
  Range (min … max):    1.752 s …  1.802 s    10 runs
 
Benchmark 2: node old/bw.js --version
  Time (mean ± σ):      2.439 s ±  0.013 s    [User: 2.587 s, System: 0.254 s]
  Range (min … max):    2.423 s …  2.466 s    10 runs
 
Summary
  node new/bw.js --version ran
    1.38 ± 0.02 times faster than node old/bw.js --version
```

## Code changes

<!--Explain the changes you've made to each file or major component. This should help the reviewer understand your changes-->
<!--Also refer to any related changes or PRs in other repositories-->

- **apps/cli/src/bw.ts:**
    - I replaced `import * as jsdom` with `import { JSDOM }`, as the CLI never uses anything else from jsdom, although it's only used for typing now.
    - I replaced the old `DOMParser` polyfill (which just immediately created a JSDOM instance) with a lazy-loaded one (which calls `new JSDOM` only when `global.DOMParser` is accessed).


## Before you submit

- Please add **unit tests** where it makes sense to do so (encouraged but not required)
- If this change requires a **documentation update** - notify the documentation team
- If this change has particular **deployment requirements** - notify the DevOps team
- Ensure that all UI additions follow [WCAG AA requirements](https://contributing.bitwarden.com/contributing/accessibility/)