@jack-weilage 2024-02-18
I've been shifting around dependencies for a while, and I've gotten the startup time ~2.6x faster. All tests are passing and no dependencies/logic have been changed.

I have some questions about the code style/normal practices for Bitwarden:

1. Is there a reason that `import type` calls are not used when something is only imported as a type?
2. Are `require()` or `import()` statements preferred when dynamically importing modules? `tsconfig.eslint.json` has `module: commonjs`, so TypeScript complains when using `import()`, but `@typescript-eslint/no-var-requires` complains when `require()` is used.

```

## Benchmark (with all changes)

Benchmark 1: node build/bw.js --version
  Time (mean ± σ):     369.4 ms ±  29.2 ms    [User: 364.7 ms, System: 59.0 ms]
  Range (min … max):   358.2 ms … 452.4 ms    10 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Benchmark 2: node build.bak/bw.js --version
  Time (mean ± σ):     968.1 ms ±  31.4 ms    [User: 1019.7 ms, System: 115.7 ms]
  Range (min … max):   949.4 ms … 1053.8 ms    10 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Summary
  node build/bw.js --version ran
    3.62 ± 0.22 times faster than node build.bak/bw.js --version
```

## Flamegraphs

**Before**
![image](https://github.com/bitwarden/clients/assets/59130272/6070bfc7-b8b1-4d49-bb89-d948364502cb)

**After**
![image](https://github.com/bitwarden/clients/assets/59130272/85b7242f-793f-434d-9f64-9eb559b1b97a)