@djsmith85 2024-02-19
Related to https://github.com/bitwarden/cli/pull/530